exercice 9 Java fini

+++ Exercices Js/exercice9/JS/traitement.js
@@ -2,12 +2,12 @@ function btnCalculer_onclick()
 { 
     var Etu, Ans65, Norm, NrbBill, Tot, TotFin, Rab;
 
-    Etu=parseFloat(document.getElementById("txtEtu").value);
-    Ans65=parseFloat(document.getElementById("txtAns65").value);
-    Norm=parseFloat(document.getElementById("txtNorm").value);
+    Etu=parseInt(document.getElementById("txtEtu").value);
+    Ans65=parseInt(document.getElementById("txtAns65").value);
+    Norm=parseInt(document.getElementById("txtNorm").value);
 
     NrbBill=Etu+Ans65+Norm;
-    Tot=(Etu*12.0)+(Ans65*15.0)+(Norm*20.0);
+    Tot=(Etu*12)+(Ans65*15)+(Norm*20);
 
     if(NrbBill>=5){
         Rab=0.1;
